tesla_fleet_rental/requirements.txt
Flask
flask-jwt-extended
requests
python-dotenv
redis
apscheduler
psutil
